Move exercises seeds to muscle_groups subdirectory
