+++ db/seeds/exercises/muscle_groups/base.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Seeds
+  module Exercises
+    module MuscleGroups
+      class Base
+        def params
+          throw 'Not implemented'
+        end
+      end
+    end
+  end
+end
