+++ db/seeds/exercises/seed.rb
@@ -1,33 +1,67 @@
 # frozen_string_literal: true
 
-require_relative 'base'
-require_relative 'abdominals'
-require_relative 'arms'
-require_relative 'back'
-require_relative 'chest'
-require_relative 'legs'
-require_relative 'shoulders'
+require_relative 'muscle_groups/base'
+require_relative 'muscle_groups/abdominals'
+require_relative 'muscle_groups/arms'
+require_relative 'muscle_groups/back'
+require_relative 'muscle_groups/chest'
+require_relative 'muscle_groups/legs'
+require_relative 'muscle_groups/shoulders'
 
 module Seeds
   module Exercises
-    class Seed < Base
+    class Seed
+      attr_reader :training_types, :muscle_groups
+
       def initialize(training_types:, muscle_groups:)
-        super(training_types: training_types, muscle_groups: muscle_groups)
+        @training_types = training_types
+        @muscle_groups  = muscle_groups
+
+        repo.clear
       end
 
-      # rubocop:disable Metrics/AbcSize
       def call
-        params = []
-        params << Abdominals.new.params
-        params << Arms.new.params
-        params << Back.new.params
-        params << Chest.new.params
-        params << Legs.new.params
-        params << Shoulders.new.params
-
-        super(params)
+        params.each do |param|
+          create_exercise(param)
+        end
       end
-      # rubocop:enable Metrics/AbcSize
+
+      private
+
+      def params
+        [
+          MuscleGroups::Abdominals.new.params,
+          MuscleGroups::Arms.new.params,
+          MuscleGroups::Back.new.params,
+          MuscleGroups::Chest.new.params,
+          MuscleGroups::Legs.new.params,
+          MuscleGroups::Shoulders.new.params
+        ]
+      end
+
+      def create_exercise(options)
+        training_type, muscle_group, items = options.values_at(:training_type, :muscle_group, :items)
+
+        items.each do |item|
+          repo.create(
+            training_type_id: training_type_id(training_type),
+            muscle_group_id: muscle_group_id(muscle_group),
+            name: item
+          )
+        end
+      end
+
+      def repo
+        @repo ||= ExerciseRepository.new
+      end
+
+      def training_type_id(training_type)
+        training_types[training_type].id
+      end
+
+      def muscle_group_id(muscle_group)
+        muscle_groups[muscle_group].id
+      end
     end
   end
 end
